feature/options-001 - python structure change
