+++ main_program.py
@@ -1,4 +1,4 @@
-from mw_srv_trade.program_schedulers.main_program import *
+from mw_srv_trade.prog_strat_ops.main_program import *
 
 scheduler_main_program_run('test', 5, 7, 3)
 #user_account_balance()
