+++ mw_srv_trade/instr_ops/__pycache__/instrument_read_write_operations.cpython-310.pyc
