+++ mw_srv_trade/mk_data_ops/data_feed_fetch.py
@@ -3,10 +3,10 @@ import datetime as dt
 from datetime import timedelta
 import pandas as pd
 from os.path import exists
-from mw_srv_trade.common_operatinos.common_ops import super_user_session
+from mw_srv_trade.comm_ops.common_ops import super_user_session
 from mw_srv_trade.constants.file_constants import *
 from mw_srv_trade.mk_data_ops.data_read_write_ops import write_data_file
-from mw_srv_trade.persistence_operations.account_management import ticks_indi, market_status
+from mw_srv_trade.persist_ops.account_management import ticks_indi, market_status
 
 from mw_srv_trade.trade_logger.logger import cus_logger
