+++ mw_srv_trade/prog_strat_ops/__pycache__/__init__.cpython-310.pyc
