+++ mw_srv_trade/prog_strat_ops/__pycache__/main_program.cpython-311.pyc
